refactor for clean architecture principals
