+++ src/main/kotlin/org/bibletranslationtools/fetcher/repository/LanguageCatalog.kt
@@ -1,4 +1,4 @@
-package org.bibletranslationtools.fetcher.domain
+package org.bibletranslationtools.fetcher.repository
 
 import org.bibletranslationtools.fetcher.data.Language
